Determine how many days in the month
Determine how many days in the month which the user enter, if the month is 1, 3, 5, 7, 8, 10, 12 then return 31 days, if the month is 4, 6, 9, 11 then return 30 days.  And then, create a method to determine whether the year is the leap year, if the year is leap year then February return 29 days, if the year isn't the leap year then return 28 days.