Print the month body
Print the month body is mean print the day of the weak.  For example, when the first day is Wednesday, then print the space to Sunday, Monday and Tuesday, and print 1 under Wednesday.  There are seven days in a weak, so print to the next line when the line has seven days.